+++ src/main/java/org/smoothbuild/compile/lang/define/DefinitionsS.java
@@ -1,11 +1,10 @@
 package org.smoothbuild.compile.lang.define;
 
 import static org.smoothbuild.util.bindings.Bindings.immutableBindings;
+import static org.smoothbuild.util.collect.Maps.concat;
 
 import org.smoothbuild.util.bindings.SingleScopeBindings;
 import org.smoothbuild.util.collect.Named;
-
-import com.google.common.collect.ImmutableMap;
 
 public record DefinitionsS(
     SingleScopeBindings<TypeDefinitionS> types,
@@ -23,11 +22,8 @@ public record DefinitionsS(
   }
 
   public <E extends Named> SingleScopeBindings<E> merge(
-      SingleScopeBindings<? extends E> outer,
-      SingleScopeBindings<? extends E> inner) {
-    var builder = ImmutableMap.<String, E>builder();
-    builder.putAll(outer.toMap());
-    builder.putAll(inner.toMap());
-    return immutableBindings(builder.build());
+      SingleScopeBindings<E> outer,
+      SingleScopeBindings<E> inner) {
+    return immutableBindings(concat(inner.toMap(), outer.toMap()));
   }
 }