+++ src/systemtest/org/smoothbuild/systemtest/slib/file/FileConstructorTest.java
@@ -5,9 +5,9 @@ import static com.google.common.truth.Truth.assertThat;
 import org.junit.jupiter.api.Test;
 import org.smoothbuild.systemtest.SystemTestCase;
 
-public class FileTest extends SystemTestCase {
+public class FileConstructorTest extends SystemTestCase {
   @Test
-  public void file_ctor() throws Exception {
+  public void file_constructor() throws Exception {
     createUserModule("""
             result = File(0x41, "name.txt");
             """);