+++ src/systemtest/org/smoothbuild/systemtest/slib/file/FileFuncTest.java
@@ -5,12 +5,12 @@ import static com.google.common.truth.Truth.assertThat;
 import org.junit.jupiter.api.Test;
 import org.smoothbuild.systemtest.SystemTestCase;
 
-public class ProjectFileTest extends SystemTestCase {
+public class FileFuncTest extends SystemTestCase {
   @Test
   public void file_from_smooth_dir_causes_error() throws Exception {
     createFile(".smooth/file.txt", "abc");
     createUserModule("""
-            result = projectFile(".smooth/file.txt");
+            result = file(".smooth/file.txt");
             """);
     runSmoothBuild("result");
     assertFinishedWithError();
@@ -21,7 +21,7 @@ public class ProjectFileTest extends SystemTestCase {
   public void file_from_smooth_subdir_causes_error() throws Exception {
     createFile(".smooth/subdir/file.txt", "abc");
     createUserModule("""
-            result = projectFile(".smooth/subdir/file.txt");
+            result = file(".smooth/subdir/file.txt");
             """);
     runSmoothBuild("result");
     assertFinishedWithError();
@@ -31,7 +31,7 @@ public class ProjectFileTest extends SystemTestCase {
   @Test
   public void illegal_path_causes_error() throws Exception {
     createUserModule("""
-            result = projectFile("..");
+            result = file("..");
             """);
     runSmoothBuild("result");
     assertFinishedWithError();
@@ -41,7 +41,7 @@ public class ProjectFileTest extends SystemTestCase {
   @Test
   public void nonexistent_path_causes_error() throws Exception {
     createUserModule("""
-            result = projectFile("nonexistent/file.txt");
+            result = file("nonexistent/file.txt");
             """);
     runSmoothBuild("result");
     assertFinishedWithError();
@@ -52,7 +52,7 @@ public class ProjectFileTest extends SystemTestCase {
   public void dir_path_causes_error() throws Exception {
     createDir("some/dir");
     createUserModule("""
-            result = projectFile("some/dir");
+            result = file("some/dir");
             """);
     runSmoothBuild("result");
     assertFinishedWithError();
@@ -63,7 +63,7 @@ public class ProjectFileTest extends SystemTestCase {
   public void file_is_returned() throws Exception {
     createFile("dir/file.txt", "abc");
     createUserModule("""
-            result = projectFile("dir/file.txt");
+            result = file("dir/file.txt");
             """);
     runSmoothBuild("result");
     assertFinishedWithSuccess();
@@ -75,7 +75,7 @@ public class ProjectFileTest extends SystemTestCase {
   public void result_is_not_cached() throws Exception {
     createFile("dir/file.txt", "abc");
     createUserModule("""
-            result = projectFile("dir/file.txt");
+            result = file("dir/file.txt");
             """);
     runSmoothBuild("result");
     assertFinishedWithSuccess();