+++ src/systemtest/org/smoothbuild/systemtest/slib/java/JunitTest.java
@@ -14,8 +14,8 @@ public class JunitTest extends SystemTestCase {
     createJunitLibs();
     createFile("src/" + SUCCESSFUL_TEST_CLASS + ".java", successfulTestSourceCode());
     createUserModule("""
-            junitJars = projectFiles("junit");
-            srcJar = projectFiles("src") > javac(libs=junitJars) > jar() > File("test.jar");
+            junitJars = files("junit");
+            srcJar = files("src") > javac(libs=junitJars) > jar() > File("test.jar");
             result = junit(tests=srcJar, deps=[]);                  
             """);
     runSmoothBuild("result");
@@ -28,8 +28,8 @@ public class JunitTest extends SystemTestCase {
     createJunitLibs();
     createFile("src/" + SUCCESSFUL_TEST_CLASS + ".java", successfulTestSourceCode());
     createUserModule("""
-            junitJars = projectFiles("junit");
-            srcJar = projectFiles("src") > javac(libs=junitJars) > jar() > File("src.jar");
+            junitJars = files("junit");
+            srcJar = files("src") > javac(libs=junitJars) > jar() > File("src.jar");
             result = junit(tests=srcJar, deps=junitJars);
             """);
     runSmoothBuild("result");
@@ -41,8 +41,8 @@ public class JunitTest extends SystemTestCase {
     createJunitLibs();
     createFile("src/" + FAILING_TEST_CLASS + ".java", failingTestSourceCode());
     createUserModule("""
-            junitJars = projectFiles("junit");
-            srcJar = projectFiles("src") > javac(libs=junitJars) > jar() > File("src.jar");
+            junitJars = files("junit");
+            srcJar = files("src") > javac(libs=junitJars) > jar() > File("src.jar");
             result = junit(tests=srcJar, deps=junitJars);
             """);
     runSmoothBuild("result");
@@ -55,8 +55,8 @@ public class JunitTest extends SystemTestCase {
     createJunitLibs();
     createDir("src");
     createUserModule("""
-            junitJars = projectFiles("junit");
-            srcJar = projectFiles("src") > javac(libs=junitJars) > jar() > File("src.jar");
+            junitJars = files("junit");
+            srcJar = files("src") > javac(libs=junitJars) > jar() > File("src.jar");
             result = junit(tests=srcJar, deps=junitJars);
             """);
     runSmoothBuild("result");
@@ -70,8 +70,8 @@ public class JunitTest extends SystemTestCase {
     createFile("src/" + SUCCESSFUL_TEST_CLASS + ".java", successfulTestSourceCode());
     createFile("src/" + FAILING_TEST_CLASS + ".java", failingTestSourceCode());
     createUserModule(format("""
-            junitJars = projectFiles("junit");
-            srcJar = projectFiles("src") > javac(libs=junitJars) > jar() > File("src.jar");
+            junitJars = files("junit");
+            srcJar = files("src") > javac(libs=junitJars) > jar() > File("src.jar");
             result = junit(include="%s", tests=srcJar, deps=junitJars);
             """, SUCCESSFUL_TEST_CLASS));
     runSmoothBuild("result");